requirements.txt
pandas
openpyxl
xlrd
glob2
webdriver-manager
selenium
pytesseract
Pillow
requests